requirements.txt
gspread==5.12.0
oauth2client==4.1.3
natsort==8.4.0
pyserial==3.5
gpiod==1.6.3 